+++ retrieval/series_cache.go
@@ -37,7 +37,7 @@ import (
 
 var (
 	droppedSeries = sidecar.OTelMeterMust.NewInt64Counter(
-		"sidecar.dropped.series",
+		config.DroppedSeriesMetric,
 		metric.WithDescription("Number of series that were dropped, not exported"),
 	)
 	keyReason = label.Key("key_reason")
@@ -390,7 +390,12 @@ func (c *seriesCache) refresh(ctx context.Context, ref uint64) error {
 		if meta == nil {
 			droppedSeriesMetadataNotFound.Add(ctx, 1)
 
-			level.Debug(c.logger).Log("msg", "metadata not found", "metric_name", metricName)
+			doevery.TimePeriod(config.DefaultNoisyLogPeriod, func() {
+				level.Warn(c.logger).Log(
+					"msg", "metadata not found",
+					"metric_name", metricName,
+				)
+			})
 			return nil
 		}
 	}